backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
openai==1.52.0
